20250930/docai-match-viewer/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pillow==10.4.0
numpy==2.1.1
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
